WebPages/nutrition4/requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fastapi==0.75.1
fonttools==4.28.5
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.28
nest-asyncio==1.5.4
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments==2.10.0
pyngrok==5.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
